Fix clippy warnings in JIT code
- Remove obsolete test code that referenced unsupported flattened fields
- Use ? operator instead of manual None check
- Remove unnecessary type casts (u32 -> u32, i64 -> i64)
- Collapse nested if statements using && operator
- Use iter().flatten() instead of manual if let Some unwrapping
- Add #[allow(clippy::missing_safety_doc)] to helpers module (safety docs in comments)

+++ facet-format-postcard/benches/vs_postcard.rs
@@ -304,7 +304,7 @@ mod vec_i32 {
         // Mix of positive and negative values to exercise ZigZag encoding
         (0..1000)
             .map(|i| {
-                let base = (i as i32) * 100;
+                let base = i * 100;
                 if i % 2 == 0 { base } else { -base }
             })
             .collect()
